chore: move shared funcs and types to common
