+++ src/common/get-namespaces.test.ts
