+++ src/common/index.ts
@@ -1,0 +1,4 @@
+export type { Ni18nOptions } from './ni18n-options'
+export type { Ni18nState } from './ni18n-state'
+export type { NamespacesNeeded } from './namespaces-needed'
+export { getNamespaces } from './get-namespaces'
