+++ src/index.ts
@@ -1,4 +1,5 @@
 export { appWithI18Next } from './app-with-i18next'
 export { loadTranslations } from './load-translations'
 export { useSyncLanguage } from './use-sync-language'
-export type { Ni18nOptions } from './ni18n-options'
+export { clientNamespaces } from './client-namespaces'
+export type { Ni18nOptions } from './common'
