+++ src/load-translations/index.ts
@@ -1,2 +1,1 @@
 export { loadTranslations } from './load-translations'
-export type { Ni18nState } from './load-translations'
